+++ examples/src/main/java/com/tenio/examples/example4/constant/Example4Constant.java
@@ -30,7 +30,6 @@ public final class Example4Constant {
   public static final int DESIGN_HEIGHT = 500;
 
   public static final int SOCKET_PORT = 8032;
-  public static final int DATAGRAM_PORT = 8034;
 
   public static final float DELAY_CREATION = 0.1f;
   // time in minutes
@@ -43,7 +42,7 @@ public final class Example4Constant {
   public static final int ONE_SECOND_EXPECT_RECEIVE_PACKETS = 10;
 
   public static final int ONE_MINUTE_EXPECT_RECEIVE_PACKETS =
-      ONE_SECOND_EXPECT_RECEIVE_PACKETS * 60 * 100;
+      ONE_SECOND_EXPECT_RECEIVE_PACKETS * 60 * NUMBER_OF_PLAYERS;
 
   private Example4Constant() {
     throw new UnsupportedOperationException();