+++ examples/src/main/java/com/tenio/examples/example6/handler/ReceivedMessageFromPlayerHandler.java
@@ -29,14 +29,14 @@ import com.tenio.common.data.ZeroMap;
 import com.tenio.common.utility.MathUtility;
 import com.tenio.core.entity.Player;
 import com.tenio.core.entity.data.ServerMessage;
-import com.tenio.core.extension.AbstractExtension;
-import com.tenio.core.extension.events.EventReceivedMessageFromPlayer;
+import com.tenio.core.handler.AbstractHandler;
+import com.tenio.core.handler.event.EventReceivedMessageFromPlayer;
 import com.tenio.examples.server.SharedEventKey;
 import java.util.ArrayList;
 import java.util.List;
 
 @Component
-public final class ReceivedMessageFromPlayerHandler extends AbstractExtension
+public final class ReceivedMessageFromPlayerHandler extends AbstractHandler
     implements EventReceivedMessageFromPlayer {
 
   @Override