+++ examples/src/main/java/com/tenio/examples/example7/handler/ReceivedMessageFromPlayerHandler.java
@@ -28,13 +28,13 @@ import com.tenio.common.bootstrap.annotation.Component;
 import com.tenio.common.data.ZeroMap;
 import com.tenio.core.entity.Player;
 import com.tenio.core.entity.data.ServerMessage;
-import com.tenio.core.extension.AbstractExtension;
-import com.tenio.core.extension.events.EventReceivedMessageFromPlayer;
+import com.tenio.core.handler.AbstractHandler;
+import com.tenio.core.handler.event.EventReceivedMessageFromPlayer;
 import com.tenio.examples.example7.constant.Example7Constant;
 import com.tenio.examples.server.SharedEventKey;
 
 @Component
-public final class ReceivedMessageFromPlayerHandler extends AbstractExtension
+public final class ReceivedMessageFromPlayerHandler extends AbstractHandler
     implements EventReceivedMessageFromPlayer {
 
   @Override