+++ framework/core/src/main/java/com/tenio/core/network/entity/session/Session.java
@@ -29,6 +29,7 @@ import com.tenio.core.entity.define.mode.PlayerDisconnectMode;
 import com.tenio.core.network.define.TransportType;
 import com.tenio.core.network.entity.packet.PacketQueue;
 import com.tenio.core.network.entity.session.manager.SessionManager;
+import com.tenio.core.network.entity.kcp.Ukcp;
 import com.tenio.core.network.zero.codec.packet.PacketReadState;
 import com.tenio.core.network.zero.codec.packet.PendingPacket;
 import com.tenio.core.network.zero.codec.packet.ProcessedPacket;
@@ -130,6 +131,30 @@ public interface Session {
   boolean containsUdp();
 
   /**
+   * Determines if the client and server can use KCP for communication. This can be used when the
+   * UDP channel is enabled on the session.
+   *
+   * @return {@code true} if KCP transportation mechanism is applied and UDP channel is using for
+   * conveying, otherwise returns {@code false}
+   */
+  boolean isEnabledKcp();
+
+  /**
+   * Allows using the KCP transportation via UDP channels.
+   *
+   * @param enabledKcp sets it {@code true} if enabled, otherwise sets it {code false}
+   */
+  void setEnabledKcp(boolean enabledKcp);
+
+  /**
+   * Determines if the session has a KCP instance for communication.
+   *
+   * @return {@code true} if the UDP channel is able to use the KCP for communication, otherwise
+   * returns {@code false}
+   */
+  boolean containsKcp();
+
+  /**
    * Retrieves a socket (TCP) channel, which is using to communicate via TCP protocol.
    *
    * @return an instance of {@link SocketChannel}
@@ -208,6 +233,20 @@ public interface Session {
    * @param remoteAddress   the {@link SocketAddress} remote address associating to the client side
    */
   void setDatagramChannel(DatagramChannel datagramChannel, SocketAddress remoteAddress);
+
+  /**
+   * Retrieves a KCP wrapper object that the session is able to use.
+   *
+   * @return a {@link Ukcp} instance
+   */
+  Ukcp getUkcp();
+
+  /**
+   * Declares a KCP wrapper object that the session is able to use.
+   *
+   * @param ukcp a {@link Ukcp} instance
+   */
+  void setUkcp(Ukcp ukcp);
 
   /**
    * Retrieves the remote address associating to the client side.